Group 5 Code v5
Added a variable gain system so different gains can be applied to different places on the tilt table. This aims to aid in balancing the ball in the center while allowing for a more exaggerated response at the edges. The XY bounds have not been tuned, nor has the PID been adjusted for the smaller area.